requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
crewai==0.1.6
dataclasses-json==0.6.3
decorator==5.1.1
emoji==2.9.0
exceptiongroup==1.2.0
executing==2.0.1
faiss-cpu==1.7.4
filetype==1.2.0
frozendict==2.3.10
frozenlist==1.4.1
geojson==2.5.0
greenlet==3.0.3
html5lib==1.1
idna==3.6
ipython==8.18.1
jedi==0.19.1
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
langchain==0.0.335
langchain-community==0.0.7
langchain-core==0.1.4
langdetect==1.0.9
langsmith==0.0.75
lxml==5.0.0
marshmallow==3.20.1
matplotlib-inline==0.1.6
multidict==6.0.4
multitasking==0.0.11
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.26.2
openai==0.28.1
packaging==23.2
pandas==2.1.4
parso==0.8.3
peewee==3.17.0
pexpect==4.9.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
pyowm==3.3.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-iso639==2023.12.11
python-magic==0.4.27
pytils==0.4.1
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.6.1
regex==2023.12.25
requests==2.31.0
sec-api==1.0.17
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.24
stack-data==0.6.3
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.5.2
tools==0.1.9
tqdm==4.66.1
traitlets==5.14.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
unstructured==0.10.25
unstructured-client==0.15.1
urllib3==2.1.0
wcwidth==0.2.12
webencodings==0.5.1
wikipedia==1.4.0
wrapt==1.16.0
yarl==1.9.4
yfinance==0.2.33
